requirements.txt
# ===========================================
# 🌐 웹 프레임워크
# ===========================================
fastapi==0.104.1
uvicorn[standard]==0.24.0

# ===========================================
# 🤖 AI/ML 라이브러리
# ===========================================
openai==1.3.8
sentence-transformers==2.2.2
torch==2.0.1
transformers==4.30.0

# 🔧 HuggingFace Hub (호환성 버전 고정)
huggingface_hub==0.16.4  # cached_download 지원 마지막 안정 버전

# ===========================================
# 🗄️ Vector Database
# ===========================================
chromadb==0.4.18

# ===========================================
# 🛠️ 유틸리티
# ===========================================
python-dotenv==1.0.0
pydantic==2.5.0
httpx==0.25.2
loguru==0.7.2
numpy==1.24.3
pandas==2.0.3

# ===========================================
# 🇰🇷 한국어 처리
# ===========================================
konlpy==0.6.0

# ===========================================
# 📊 AI Hub 데이터 처리 (추가)
# ===========================================
# JSON 스트리밍 처리
ijson==3.2.3

# 데이터 분석 및 통계
scipy==1.11.4
scikit-learn==1.3.2

# 텍스트 전처리 및 정규식
regex==2023.10.3

# 날짜/시간 처리
python-dateutil==2.8.2

# ===========================================
# 🔧 개발 및 테스팅 (Docker에서 제외 가능)
# ===========================================
# 주석 처리하여 빌드 시간 단축
# pytest==7.4.3
# pytest-asyncio==0.21.1

# 성능 모니터링 (프로덕션에서만 필요시)
# psutil==5.9.6

# ===========================================
# 🐳 Docker 최적화
# ===========================================
# Gunicorn (프로덕션 배포용)
gunicorn==22.0.0

# ===========================================
# 📝 추가 유틸리티
# ===========================================
# 환경 변수 관리
python-decouple==3.8

# 참고: uuid, time은 Python 표준 라이브러리 사용